Add rudimentary solver.
Heavily inspired by my Fill-a-Pix solver from a year ago:
https://github.com/orez-/fillapix-solver/blob/572d8c1c4d93a960e30b3f723801d4dac612e809/fillapix.py#L136

Unfortunately, while Fill-a-Pix is a game of no hidden information,
Hexcells is. This solve function will return everything it can figure
out with the information given, including building off of what it
figures out, but it cannot solve the entire puzzle given a starting
configuration. Next step: having it enter its results in-game, take a
screenshot, and repeat the process.