Use numpy for hex identification.
Holy bajeez numpy is like legitimately over sixty times faster than
whatever I was doing. I'd love to get it working with the OCR to
really complete the numpy sandwich.

This shake-up has also brought some issues to light. Since the
`find_objects` function seems to scan from the top, and our hex
positioning function takes the first hex it finds as the origin and
positions subsequent hexes relative to that, some of my test cases
are parsing incorrectly due to rounding errors. I've updated the
hardcoded spacing multiplier for now (yuck), but we might need a better
solution soon. If we have the guarantee of the top-to-bottom iteration,
maybe we measure each hexagon from the previous hex, or maybe the
closest hex. Maybe we just smooth the multiplier as we place more
hexagons.

Also even after updating the multiplier, the solver for that test case
is double-placing. I haven't looked into that at all yet.